+++ @remirror/cli/jest.config.js
@@ -1,0 +1,8 @@
+const config = require('../../support/jest/jest.config');
+
+module.exports = {
+  ...config,
+  name: '@remirror/cli',
+  displayName: 'cli',
+  testEnvironment: 'node',
+};