+++ @remirror/cli/src/commands/bundle/types.ts
@@ -1,0 +1,6 @@
+export interface BundleArgv {
+  /**
+   * The file path of the js / ts file to bundle together
+   */
+  source: string;
+}