+++ @remirror/native-bridge/jest.config.js
@@ -1,0 +1,7 @@
+const config = require('../../support/jest/jest.config');
+
+module.exports = {
+  ...config,
+  name: '@remirror/native-bridge',
+  displayName: 'native-bridge',
+};