+++ examples/with-react-native/index.js
@@ -1,0 +1,5 @@
+import { AppRegistry } from 'react-native';
+import Entry from './src/entry';
+import { name as appName } from './app.json';
+
+AppRegistry.registerComponent(appName, () => Entry);