+++ examples/with-react-native/src/screens/ui-twitter.tsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import { Text, View } from 'react-native';
+
+export class UITwitterScreen extends Component {
+  public render() {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <Text>UI Twitter</Text>
+      </View>
+    );
+  }
+}